refactor: Share core intialization from auth and app
