+++ src/script/service/CoreSingleton.ts
@@ -18,12 +18,24 @@
  */
 
 import {Account} from '@wireapp/core';
+import {ClientType} from '@wireapp/api-client/src/client';
 import {container, singleton} from 'tsyringe';
 import {APIClient} from './APIClientSingleton';
+import {createStorageEngine, DatabaseTypes} from './StoreEngineProvider';
+
+const clientToDbType: Record<ClientType, DatabaseTypes> = {
+  [ClientType.PERMANENT]: DatabaseTypes.PERMANENT,
+  [ClientType.TEMPORARY]: DatabaseTypes.TEMPORARY,
+  [ClientType.NONE]: DatabaseTypes.EFFEMERAL,
+};
 
 @singleton()
 export class Core extends Account {
   constructor(apiClient = container.resolve(APIClient)) {
-    super(apiClient);
+    super(apiClient, {
+      createStore: (storeName, context) => {
+        return createStorageEngine(storeName, clientToDbType[context.clientType]);
+      },
+    });
   }
 }
